+++ CHANGELOG.rst
@@ -1,3 +1,11 @@
+version 2.5.0
+----------------------------------------------------------
+* Support response types `stray` and `empty` in repl server.
+* Minor updates in asyncio server.
+* Update reactive server to send stray response of given length.
+* Transaction manager updates on retries for empty and invalid packets.
+* Test fixes for asyncio client and transaction manager.
+
 version 2.5.0rc3
 ----------------------------------------------------------
 * Minor fix in documentations